+++ app/src/services/LabelServices.ts
@@ -4,6 +4,7 @@
  */
 export const labelServices = {
     getLabels,
+    setLabels
 }
 
 function getLabels(project_name: any) {
@@ -21,6 +22,31 @@ function getLabels(project_name: any) {
         })
  }
 
+function setLabels(project: string, label_name: string, firebase:any) {
+    const requestOptions = {
+        method: 'POST',
+         headers: { 'Content-Type': 'application/json' },
+         body: JSON.stringify({ label_name })
+     };
+
+     const token = localStorage.getItem('user-token');
+     if (firebase.auth.currentUser != null) {
+         firebase.auth.currentUser.getIdToken().then((idToken: string) => {
+             if (token !== idToken) {
+                 localStorage.setItem('user-token', idToken)
+             }
+         })
+     } else {
+         window.location.href = '/auth';
+     }
+
+    return fetch(process.env.REACT_APP_API_URL +
+        '/projects/' + project + '/labels/add' + '?id_token=' + localStorage.getItem('user-token'), requestOptions)
+        .then(handleResponse)
+        .then(data => {
+            return data.users
+        })
+ }
 function handleResponse(response: { text: () => Promise<any>; ok: any; status: number; statusText: any; }) {
    return response.text().then((text: string) => {
        const data = text && JSON.parse(text);
