+++ app/src/services/ProjectServices.ts
@@ -9,7 +9,6 @@ export const projectServices = {
     getProjectUsers,
     setProjectUsers,
     createProject,
-    getProjectTags,
     setProjectTags
 }
 
@@ -117,34 +116,6 @@ async function getProjectUsers(project: string, firebase: any) {
         .then(handleResponse)
         .then(data => {
             return data.users
-        })
- }
-
- async function getProjectTags(project: string, firebase:any) {
-    const requestOptions = {
-        method: 'GET',
-        headers: { 'Content-Type': 'application/json',
-        "Access-Control-Allow-Origin": "*",
-        "Access-Control-Allow-Methods": "DELETE, POST, GET, OPTIONS",
-        "Access-Control-Allow-Headers": "Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With" },
-    };
-
-    const token = localStorage.getItem('user-token');
-     if (firebase.auth.currentUser != null) {
-         firebase.auth.currentUser.getIdToken().then((idToken: string) => {
-             if (token !== idToken) {
-                 localStorage.setItem('user-token', idToken)
-             }
-         })
-     } else {
-         window.location.href = '/auth';
-     }
-
-    return fetch(process.env.REACT_APP_API_URL +
-        '/projects/' + project + '/labels/all' + '?id_token=' + localStorage.getItem('user-token'), requestOptions)
-        .then(handleResponse)
-        .then(data => {
-            return data.labels
         })
  }
